+++ company-clang.el
@@ -48,6 +48,13 @@ and `c-electric-colon', for automatic completion right after \">\" and
 \":\"."
   :type 'boolean)
 
+(defcustom company-clang-use-compile-flags-txt nil
+  "When non-nil, use flags from compile_flags.txt if present.
+
+The lines from that files will be appended to `company-clang-arguments'."
+  :type 'boolean
+  :safe 'booleanp)
+
 (defcustom company-clang-arguments nil
   "Additional arguments to pass to clang when completing.
 Prefix files (-include ...) can be selected with `company-clang-set-prefix'
@@ -249,12 +256,30 @@ or automatically through a custom `company-clang-prefix-guesser'."
   (append '("-fsyntax-only" "-Xclang" "-code-completion-macros")
           (unless (company-clang--auto-save-p)
             (list "-x" (company-clang--lang-option)))
-          company-clang-arguments
+          (company-clang--arguments)
           (when (stringp company-clang--prefix)
             (list "-include" (expand-file-name company-clang--prefix)))
           (list "-Xclang" (format "-code-completion-at=%s"
                                   (company-clang--build-location pos)))
           (list (if (company-clang--auto-save-p) buffer-file-name "-"))))
+
+(defun company-clang--arguments ()
+  (let ((fname "compile_flags.txt")
+        (args company-clang-arguments))
+    (when company-clang-use-compile-flags-txt
+      (let ((dir (locate-dominating-file default-directory fname)))
+        (when dir
+          (with-temp-buffer
+            (insert-file-contents (expand-file-name fname dir))
+            (setq args
+                  (append
+                   args
+                   (split-string (buffer-substring-no-properties
+                                  (point-min) (point-max))
+                                 "[\n\r]+"
+                                 t
+                                 "[ \t]+")))))))
+    args))
 
 (defun company-clang--candidates (prefix callback)
   (and (company-clang--auto-save-p)