+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+* New user option `company-clang-use-compile-flags-txt`
+  ([#933](https://github.com/company-mode/company-mode/issues/933)).
 * Support for completion style specific sorting (Emacs 27 feature).
 * Snippet/template field interaction is inhibited while completion is active
   (where by default `TAB` calls `company-complete-common`, clashing with snippet